elevia-scheduler/requirements.txt
fastapi
uvicorn
google-api-python-client
google-auth
google-auth-oauthlib
pydantic
python-dotenv
